Moved telecommands implementation to separate library
