+++ libs/obc/communication/CMakeLists.txt
@@ -11,7 +11,11 @@ target_link_libraries(${NAME}
 	comm
 	telecommunication
 	settings
+	obc_telecommands
 )
 
 target_include_directories(${NAME} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/Include)
 target_include_directories(${NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Include/obc)
+
+
+add_subdirectory(telecommands)
