+++ libs/obc/communication/communication.cpp
@@ -6,31 +6,9 @@
 using std::uint8_t;
 using gsl::span;
 using drivers::i2c::II2CBus;
-using devices::comm::ITransmitFrame;
-using obc::PingTelecommand;
 using telecommunication::uplink::IHandleTeleCommand;
-using telecommunication::downlink::DownlinkFrame;
-using telecommunication::downlink::DownlinkAPID;
 
 using namespace obc;
-
-void PingTelecommand::Handle(ITransmitFrame& transmitter, span<const uint8_t> parameters)
-{
-    UNREFERENCED_PARAMETER(parameters);
-
-    const char* response = "PONG";
-
-    DownlinkFrame frame(DownlinkAPID::Pong, 0);
-
-    frame.PayloadWriter().WriteArray(gsl::make_span(reinterpret_cast<const uint8_t*>(response), 4));
-
-    transmitter.SendFrame(frame.Frame());
-}
-
-uint8_t PingTelecommand::CommandCode() const
-{
-    return static_cast<uint8_t>('P');
-}
 
 Telecommands::Telecommands()
     : _ping(), //
