+++ libs/obc/communication/telecommands/CMakeLists.txt
@@ -1,0 +1,15 @@
+set(NAME obc_telecommands)
+
+set(SOURCES
+    ping.cpp
+)
+
+add_library(${NAME} STATIC ${SOURCES})
+
+target_link_libraries(${NAME} 
+	base
+	telecommunication
+)
+
+target_include_directories(${NAME} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/Include)
+target_include_directories(${NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Include/obc/telecommands)
